+++ backend/api/src/main/java/com/yat2/episode/episode/EpisodeController.java
@@ -22,9 +22,9 @@ import java.util.List;
 import java.util.UUID;
 
 import com.yat2.episode.episode.dto.EpisodeDetail;
-import com.yat2.episode.episode.dto.EpisodeSearchReq;
-import com.yat2.episode.episode.dto.MindmapEpisodeRes;
-import com.yat2.episode.episode.dto.StarUpdateReq;
+import com.yat2.episode.episode.dto.request.EpisodeSearchReq;
+import com.yat2.episode.episode.dto.request.StarUpdateReq;
+import com.yat2.episode.episode.dto.response.MindmapEpisodeRes;
 import com.yat2.episode.global.exception.ErrorCode;
 import com.yat2.episode.global.swagger.ApiErrorCodes;
 import com.yat2.episode.global.swagger.AuthRequiredErrors;
