+++ backend/api/src/main/java/com/yat2/episode/episode/EpisodeRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 import java.util.UUID;
 
-import com.yat2.episode.episode.dto.EpisodeSummaryRes;
+import com.yat2.episode.episode.dto.response.EpisodeSummaryRes;
 
 @Repository
 public interface EpisodeRepository extends JpaRepository<Episode, UUID> {
