+++ backend/api/src/main/java/com/yat2/episode/episode/dto/request/EpisodeSearchReq.java
@@ -1,4 +1,4 @@
-package com.yat2.episode.episode.dto;
+package com.yat2.episode.episode.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
