+++ backend/api/src/main/java/com/yat2/episode/episode/dto/response/MindmapEpisodeRes.java
@@ -1,7 +1,9 @@
-package com.yat2.episode.episode.dto;
+package com.yat2.episode.episode.dto.response;
 
 import java.util.List;
 import java.util.UUID;
+
+import com.yat2.episode.episode.dto.EpisodeDetail;
 
 public record MindmapEpisodeRes(
         UUID mindmapId,
