+++ backend/api/src/test/java/com/yat2/episode/utils/TestEntityFactory.java
@@ -11,8 +11,8 @@ import java.util.UUID;
 
 import com.yat2.episode.episode.Episode;
 import com.yat2.episode.episode.EpisodeStar;
-import com.yat2.episode.episode.dto.EpisodeUpsertContentReq;
-import com.yat2.episode.episode.dto.StarUpdateReq;
+import com.yat2.episode.episode.dto.request.EpisodeUpsertContentReq;
+import com.yat2.episode.episode.dto.request.StarUpdateReq;
 import com.yat2.episode.mindmap.Mindmap;
 import com.yat2.episode.mindmap.MindmapParticipant;
 import com.yat2.episode.user.User;
